test optimization
I generated an encryption key using the Fernet.generate_key() method and saved it in a file called "chave.key". This key will be used both to encrypt and to describe the files.

I created a list of named files that hold the names of the files that will be encrypted. I used a list comprehension to filter out unwanted files (such as the script itself and the key file) from the result returned by os.listdir().

I used a loop to go through each file in the file list and perform the encryption. Open each file in binary ("rb") read mode, read the contents and store them in the content variable.

I created a new variable calledcrypt_content to store the encrypted content, getting the Fernet.encrypt() passing the key and the content.

I re-opened the same file in binary writing ("wb") mode and wrote the encrypted content into the file.